+++ backend/sql/queries/notes.sql
@@ -1,4 +1,4 @@
--- name: NewNote :one
+-- name: NewNote :exec
 INSERT INTO notes (id, created_at, updated_at, body, user_id)
 VALUES (
     gen_random_uuid (),
@@ -6,17 +6,16 @@ VALUES (
     NOW(),
     $1,
     $2 
-)
-RETURNING *;
+);
 
 -- name: GetAllNotes :many
-SELECT * FROM notes ORDER BY created_at ASC;
+SELECT * FROM notes WHERE user_id = $1 ORDER BY created_at ASC ;
 
 -- name: GetNoteByID :one
-SELECT * FROM notes WHERE id = $1;
+SELECT * FROM notes WHERE id = $1 AND user_id = $2;
 
 -- name: DeleteNote :exec
-DELETE FROM notes WHERE id = $1;
+DELETE FROM notes WHERE id = $1 AND user_id = $2;
 
 -- name: UpdateNote :exec
 UPDATE notes
@@ -24,7 +23,4 @@ SET
     updated_at = NOW(),
     body = $1
 WHERE 
-    id = $2;
-
--- name: GetNotesByAuthor :many
-SELECT * FROM notes WHERE user_id = $1 ORDER BY created_at ASC;
+    id = $2 AND user_id = $2;
